rename folder providers, get API should return an empty value
