+++ master/buildbot/secrets/providers/__init__.py
