+++ master/buildbot/secrets/providers/base.py
